bug与优化
